# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * ComicVerse API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Character {
  id: number;
  userId: string;
  name: string;
  description: string;
  personality: string;
  visualStyle: string;
  avatarUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateCharacterInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name: string;
  description: string;
  personality: string;
  visualStyle: string;
  avatarUrl?: string | null;
}

export interface Comic {
  id: number;
  userId: string;
  title: string;
  description?: string | null;
  style: string;
  template: string;
  published: boolean;
  coverImageData?: string | null;
  likesCount: number;
  userHasLiked: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface Panel {
  id: number;
  comicId: number;
  order: number;
  dialogue?: string | null;
  caption?: string | null;
  imageData?: string | null;
  imagePrompt?: string | null;
  characterIds: number[];
  createdAt: string;
  updatedAt: string;
}

export type ComicWithPanels = Comic & {
  panels: Panel[];
};

export interface CreateComicInput {
  /**
     * @minLength 1
     * @maxLength 200
     */
  title: string;
  description?: string | null;
  style: string;
  template: string;
}

export interface UpdateComicInput {
  title?: string;
  description?: string | null;
  style?: string;
  template?: string;
  coverImageData?: string | null;
}

export interface CreatePanelInput {
  order?: number;
  dialogue?: string | null;
  caption?: string | null;
  imageData?: string | null;
  imagePrompt?: string | null;
  characterIds?: number[];
}

export interface UpdatePanelInput {
  order?: number;
  dialogue?: string | null;
  caption?: string | null;
  imageData?: string | null;
  imagePrompt?: string | null;
  characterIds?: number[];
}

export interface GenerateStoryInput {
  prompt: string;
  style: string;
  panelCount?: number;
  characterNames?: string[];
}

export type GeneratedStoryPanelsItem = {
  order: number;
  caption: string;
  dialogue: string;
  imagePrompt: string;
};

export interface GeneratedStory {
  title: string;
  description: string;
  panels: GeneratedStoryPanelsItem[];
}

export interface GenerateDialogueInput {
  panelDescription: string;
  characterNames?: string[];
  style: string;
  context?: string;
}

export interface GeneratedDialogue {
  dialogue: string;
  caption: string;
}

export interface GenerateImageInput {
  prompt: string;
  style: string;
  panelDescription?: string;
}

export interface GeneratedImage {
  /** base64 encoded PNG */
  imageData: string;
}

export interface CommunityComicSummary {
  id: number;
  title: string;
  description?: string | null;
  style: string;
  template: string;
  coverImageData?: string | null;
  likesCount: number;
  userHasLiked: boolean;
  authorName: string;
  authorId: string;
  panelCount: number;
  createdAt: string;
}

export interface CommunityFeed {
  comics: CommunityComicSummary[];
  total: number;
  page: number;
  totalPages: number;
}

export type CommunityComicDetail = CommunityComicSummary & {
  panels: Panel[];
};

export interface LikeStatus {
  liked: boolean;
  likesCount: number;
}

export type DashboardStatsRecentActivityItemType = typeof DashboardStatsRecentActivityItemType[keyof typeof DashboardStatsRecentActivityItemType];


export const DashboardStatsRecentActivityItemType = {
  comic_created: 'comic_created',
  comic_published: 'comic_published',
  like_received: 'like_received',
} as const;

export type DashboardStatsRecentActivityItem = {
  type: DashboardStatsRecentActivityItemType;
  comicId: number;
  comicTitle: string;
  timestamp: string;
};

export interface DashboardStats {
  totalComics: number;
  publishedComics: number;
  totalLikesReceived: number;
  totalCharacters: number;
  recentActivity: DashboardStatsRecentActivityItem[];
}

/**
 * Resource not found
 */
export type NotFoundResponse = ErrorResponse;

export type PublishComicBody = {
  published: boolean;
};

export type GetCommunityFeedParams = {
page?: number;
limit?: number;
style?: string;
};