# lib/api-zod/src/generated/types/comicWithPanels.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * ComicVerse API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Comic } from './comic';
import type { Panel } from './panel';

export type ComicWithPanels = Comic & {
  panels: Panel[];
};