# lib/api-zod/src/generated/types/panel.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * ComicVerse API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Panel {
  id: number;
  comicId: number;
  order: number;
  dialogue?: string | null;
  caption?: string | null;
  imageData?: string | null;
  imagePrompt?: string | null;
  characterIds: number[];
  createdAt: Date;
  updatedAt: Date;
}